Move get requests and tests to separate files
