+++ internals/handlers/flights/get_flights.go
@@ -10,6 +10,33 @@ import (
 	"github.com/rpparas/flight_log/internals/model"
 )
 
+// GetFlight func one flight by ID
+// @Description Get one flight by ID
+// @Tags Flight
+// @Accept json
+// @Produce json
+// @Success 200 {object} model.Flight
+// @router /api/v1/flights/{id} [get]
+func GetFlight(c *fiber.Ctx) error {
+	db := database.DB
+	var flights model.Flight
+
+	// Read the param flightsId
+	id := c.Params("flightsId")
+
+	// Find the flights with the given Id
+	db.Find(&flights, "id = ?", id)
+
+	// If no such flights present return an error
+	if flights.ID == uuid.Nil {
+		// TODO: Determine appropriate error code, 204 or 404 instead of 200
+		return c.Status(200).JSON(fiber.Map{"status": "no content", "message": "No flight found", "data": nil})
+	}
+
+	// Return the flights with the Id
+	return c.JSON(fiber.Map{"status": "success", "message": "Flights Found", "data": flights})
+}
+
 // GetFlights func gets all existing flights
 // @Description Get all existing flights
 // @Tags Flights
@@ -152,59 +179,4 @@ func parseQuerymaxDurationMins(c *fiber.Ctx) int {
 		return -1
 	}
 	return intVal
-}
-
-// CreateFlight func create a flight
-// @Description Create a flight
-// @Tags Flights
-// @Accept json
-// @Produce json
-// @Success 200 {object} model.Flight
-// @router /api/v1/flights [post]
-func CreateFlight(c *fiber.Ctx) error {
-	db := database.DB
-	flights := new(model.Flight)
-
-	// Store the body in the flights and return error if encountered
-	err := c.BodyParser(flights)
-	if err != nil {
-		return c.Status(500).JSON(fiber.Map{"status": "error", "message": "Review your input", "data": err})
-	}
-	// Add a uuid to the flights
-	flights.ID = uuid.New()
-	// Create the Flight and return error if encountered
-	err = db.Create(&flights).Error
-	if err != nil {
-		return c.Status(500).JSON(fiber.Map{"status": "error", "message": "Could not create flights", "data": err})
-	}
-
-	// Return the created flights
-	return c.JSON(fiber.Map{"status": "success", "message": "Created Flight", "data": flights})
-}
-
-// GetFlight func one flight by ID
-// @Description Get one flight by ID
-// @Tags Flight
-// @Accept json
-// @Produce json
-// @Success 200 {object} model.Flight
-// @router /api/v1/flights/{id} [get]
-func GetFlight(c *fiber.Ctx) error {
-	db := database.DB
-	var flights model.Flight
-
-	// Read the param flightsId
-	id := c.Params("flightsId")
-
-	// Find the flights with the given Id
-	db.Find(&flights, "id = ?", id)
-
-	// If no such flights present return an error
-	if flights.ID == uuid.Nil {
-		// TODO: Determine appropriate error code, 204 or 404 instead of 200
-		return c.Status(200).JSON(fiber.Map{"status": "no content", "message": "No flight found", "data": nil})
-	}
-
-	// Return the flights with the Id
-	return c.JSON(fiber.Map{"status": "success", "message": "Flights Found", "data": flights})
 }
